feat: Add dashboard components and types for video handling
- Introduced new types for video cards, user preferences, and streaming progress.
- Implemented EmptyState, LoadingModal, PreferenceSetup, StreamingProgress, UrlInputSection, VideoCard, VideoCarousel, VideoOptionsModal components.
- Added utility functions for formatting view counts, converting video data, parsing streaming messages, and scrolling sliders.
