+++ elcondorpasa/src/components/dashboard/EmptyState.tsx
@@ -1,0 +1,19 @@
+import { LucideIcon } from "lucide-react";
+
+interface EmptyStateProps {
+  icon: LucideIcon;
+  title: string;
+  description: string;
+}
+
+export const EmptyState: React.FC<EmptyStateProps> = ({
+  icon: Icon,
+  title,
+  description,
+}) => (
+  <div className="bg-[#2A2A2A] rounded-2xl p-8 text-center">
+    <Icon className="w-16 h-16 text-gray-600 mx-auto mb-4" />
+    <h3 className="text-xl font-semibold mb-2">{title}</h3>
+    <p className="text-gray-400">{description}</p>
+  </div>
+);
